Update User Interface
In the styles.css relevant css was used so the form would be
aligned to the bottom of the screen and centered. Additionally,
the chat container after a prompt was entered would be lcoated
in the center of the screen.

In the script.js, javascript was used to allow for text to be
submitted to the chat form, and a response from the robot would
output. A function was also added so that if the enter key was
pressed, a prompt would be submitted.