Update Chat Prompt Form Position
The index.html file was updated so it would initially have
the app id and centered class. The centered class would center
the chat prompt form both horizontally and vertically on the screen.

After a prompt is entered, the class is switched to active. This is
done through javascript in the handleSubmit function.

The styles.css was updated so the styling would reflect this.