Add Light/Dark modified
The styles.css and script.js file were modified to allow for
toggling between light and dark mode.

This involved relevant styling to the styles.css and
an eventlistener function themeToggleBtn that is linked to the html
button id. Additionally, since this applies to the chatstripe
icons too, the source was set to a nested ternary operation to choose
between light and dark mode icons.