Update script.js
The client side script.js handleSubmit function was updated so it
would be able to fetch the AI response from the API and then format
it properly in the chat container. Some error checking was also
included in case the fetching failed.

Some minor changes were also done to the chatStripe function for
rendering the light/dark mode icons.